project/requirements.txt
torch>=1.10.0
torchvision>=0.11.1
transformers>=4.15.0
matplotlib>=3.4.3
numpy>=1.21.0
Pillow>=8.3.1
tqdm>=4.62.3
scikit-learn>=0.24.2
albumentations>=1.0.3
opencv-python>=4.5.3
pandas>=1.3.3
tensorboard>=2.7.0
timm>=0.4.12
PyYAML>=6.0
nltk>=3.6.5 